Add emergency stop and verifier router contracts for version management (#80)
As we get ready to deploy RISC Zero verifiers to Ethewreum mainnet, we
need a system to manage versions of the verifier contract and proof
system. In particular, we need a way to respond to critical bugs, and we
need a way to make upgraded versions of the proof system available to
developers. Both mechanisms need to be opt-in so that applications can
choose the solution that works best for their security model.

This PR implements a basic version management system, based around two
new contracts:

* `RiscZeroVerifierEmergencyStop` is a wrapper around a verifier
contract that allows for the verifier to be shutdown if a critical bug
is discovered.
* `RiscZeroVerifierRouter` is a wrapper around a list of verifiers that
allows new verifiers to be added to support proof system upgrades.

These are intended to be used together with an authorized multisig, and
a `TimelockController` on addition of new verifiers.

This PR is currently marked as draft, as the verifier identifiers are
not yet fully implemented in the `risc0` codebase.

Also addresses https://github.com/risc0/risc0-ethereum/issues/30 as part
of a refactor to the `IRiscZeroVerifier` interface.

Related to https://github.com/risc0/risc0/pull/1794, from which the
verifier parameters have been bootstrapped.

---------

Co-authored-by: Paul <paul.gafni@gmail.com>