feat: efficient multi view call support (#91)
Proposal to close #90 

The least changing API would be to switch the `preflight` and `execute`
methods to take in a mutable reference to the `env`, but this comes with
the downside of having a breaking change from previous APIs (unless
using different names) and an awkward API imo.

This simple change mentioned above would look sth like:

```rust
 	// Host
    let mut env = EthViewCallEnv::from_provider(provider, BLOCK)?;
    let _result = ViewCall::new(call, contract).preflight(&mut env)?;
    let _result2 = ViewCall::new(call2, contract).preflight(&mut env)?;
    let input = env.into_zkvm_input()?

	// Guest
    let env = input.into_env();
    let result = ViewCall::new(call, contract).execute(&env);
 	let result2 = ViewCall::new(call2, contract).execute(&env);
```

Where the proposal in this API flips the env and `ViewCall`, as it's
more clear the call is operating on the `env`:

```rust
 	// Host
    let env = EthViewCallEnv::from_provider(provider, BLOCK)?;
    let _result = env.preflight(ViewCall::new(call, contract))?;
	let _result2 = env.preflight(ViewCall::new(call2, contract))?;
    let input = env.into_zkvm_input()?

	// Guest
    let env = input.into_env();
    let result = env.execute(ViewCall::new(call, contract));
 	let result2 = env.execute(ViewCall::new(call2, contract));
```

~~Opening as draft until approval, and I will add tests and an example
with this.~~ basic test added and example could be #95, although that
example probably isn't practical without historical queries supported
better, in the example or in steel

Another option is to create APIs around a type like `MultiViewCall`, and
describe the multiple calls that way which returns a tuple of results
with the zkvm input. I didn't love this option, as it would just be
confusing to use, but figured I'd mention that also. Happy to elaborate
on this if helpful, though!

---------

Co-authored-by: Wolfgang Welz <welzwo@gmail.com>