Use the `main` branch of the `risc0` monorepo for development (#93)
This changes all the references for packages from `risc0` in the
`Cargo.toml` files. During development, we want to stay up to date with
the `main` branch of `risc0`, and only switch to a stable release during
the release process. With `1.0` approaching very soon now, the rate of
breaking changes should slow significantly.

Resolves: https://github.com/risc0/risc0-ethereum/issues/73
Resolves: https://github.com/risc0/risc0-ethereum/issues/89

Co-authored-by: carlosgj94

---------

Co-authored-by: capossele <angelocapossele@gmail.com>