Implement call builder pattern to more closely match alloy APIs (#115)
More context here
https://github.com/risc0/risc0-ethereum/pull/91#issuecomment-2100548707

This is built on #110 

Draft until:
- [x] Update examples and usages to new API
- [x] Add docs
- [x] Deprecate `ViewCall` constructor
- [x] Update old deprecation warning to point to this API
- [x] Remove function added in pull 91, as that was between released
versions (keep if this comes in after release)

Although I think this API is better for now and to future proof, I don't
feel strongly about any names or specific APIs. My goal was to try to
match alloy as close as possible, while keeping necessary patterns or
naming that existed before (also to minimize breaking changes like
re-using `ViewCall`. Opinions or comments welcome!

This _should_ only give deprecation warnings from 0.10

Details that might be missed:
- Generic bounds loosened at type level (where it made sense to)
- `new` and `preflight` are the constructors for `Contract`. They both
have the `call` method at the end of the chain, but the guest one
implicitly unwraps the error. I don't love this pattern, but I'm
matching the existing one
- Happy to give argument of why we maybe should have it both return a
result
- Contract and view call types moved into `contract.rs`
- `SteelEnv` trait the internal implementation is based on is kept
pseudo-public, to avoid exposing the trait type to the public API. Just
used to be able to use `CallBuilder` in a generic way internally.
- This doesn't lend the APIs well for people to create their custom
implementations outside of our supported ones. Intentionally so that the
API surface we have to maintain is smaller, at least until someone asks
for this.

---------

Co-authored-by: Victor Graf <victor@risczero.com>
Co-authored-by: capossele <angelocapossele@gmail.com>
Co-authored-by: Wolfgang Welz <welzwo@gmail.com>