WEB3-534: Add config ID check to Steel.sol and make changes required in implementation of PoVW (#17)
This PR includes a new check for the config ID in Steel.sol. When a guest suports queries to exactly one chain, and includes the config ID as a constant, this check is not required and so we have not previously implemented it. However, if a single guest binary is intended to potentially query multiple chains (either in the same of distinct executions), the config ID must be checked to bind the EVM behavior such as the hardforks applies and the chain ID reported by block.chainid.

PoVW makes use of this new functionality. It also provides some extensions upon the base EvmEnv for easier handling of a sequence of blocks. This PR includes the changes required to support PoVW.